--- src/main/java/com/wanted/codebombalms/domain/problem/category/dto/request/CategoryRequest.java
@@ -1,4 +1,0 @@
-package com.wanted.codebombalms.domain.problem.category.dto.request;
-
-public class CategoryRequest {
-}
