--- src/main/java/com/wanted/codebombalms/domain/problem/category/exception/CategoryNotFoundException.java
@@ -1,4 +1,0 @@
-package com.wanted.codebombalms.domain.problem.category.exception;
-
-public class CategoryNotFoundException {
-}
