--- src/main/java/com/wanted/codebombalms/domain/problem/problem/dto/request/ProblemRequest.java
@@ -1,4 +1,0 @@
-package com.wanted.codebombalms.domain.problem.problem.dto.request;
-
-public class ProblemRequest {
-}
