--- src/main/java/com/wanted/codebombalms/domain/problem/problem/exception/ProblemNotFound.java
@@ -1,4 +1,0 @@
-package com.wanted.codebombalms.domain.problem.problem.exception;
-
-public class ProblemNotFound {
-}
