--- src/main/java/com/wanted/codebombalms/domain/problem/set/dto/request/SetRequest.java
@@ -1,4 +1,0 @@
-package com.wanted.codebombalms.domain.problem.set.dto.request;
-
-public class SetRequest {
-}
