--- src/main/java/com/wanted/codebombalms/domain/problem/set/exception/SetNotFoundException.java
@@ -1,8 +1,0 @@
-package com.wanted.codebombalms.domain.problem.set.exception;
-
-public class SetNotFoundException extends RuntimeException {
-
-    public SetNotFoundException(String message) {
-        super(message);
-    }
-}
