+++ src/main/java/com/wanted/codebombalms/domain/problems/category/dto/request/CategoryRequest.java
@@ -1,0 +1,4 @@
+package com.wanted.codebombalms.domain.problems.category.dto.request;
+
+public class CategoryRequest {
+}
