+++ src/main/java/com/wanted/codebombalms/domain/problems/category/exception/CategoryNotFoundException.java
@@ -1,0 +1,4 @@
+package com.wanted.codebombalms.domain.problems.category.exception;
+
+public class CategoryNotFoundException {
+}
