+++ src/main/java/com/wanted/codebombalms/domain/problems/problem/dto/request/ProblemRequest.java
@@ -1,0 +1,4 @@
+package com.wanted.codebombalms.domain.problems.problem.dto.request;
+
+public class ProblemRequest {
+}
