+++ src/main/java/com/wanted/codebombalms/domain/problems/problem/exception/ProblemNotFound.java
@@ -1,0 +1,4 @@
+package com.wanted.codebombalms.domain.problems.problem.exception;
+
+public class ProblemNotFound {
+}
