+++ src/main/java/com/wanted/codebombalms/domain/problems/set/dto/request/SetRequest.java
@@ -1,0 +1,4 @@
+package com.wanted.codebombalms.domain.problems.set.dto.request;
+
+public class SetRequest {
+}
