+++ src/main/java/com/wanted/codebombalms/domain/problems/set/entity/ProblemSet.java
@@ -1,0 +1,40 @@
+package com.wanted.codebombalms.domain.problems.set.entity;
+
+import com.wanted.codebombalms.domain.problems.category.entity.ProblemCategory;
+import jakarta.persistence.*;
+
+@Entity
+public class ProblemSet {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long problemSetId;
+
+    @ManyToOne
+    @JoinColumn(name = "category_id")
+    private ProblemCategory category;
+
+    @Column(nullable = false)
+    private String title;
+
+    private String description;
+
+    private String difficulty;
+
+    @Column(nullable = false)
+    private String status;
+
+    protected ProblemSet() {
+    }
+
+    public Long getProblemSetId() {
+        return problemSetId;
+    }
+
+    public ProblemCategory getCategory() {
+        return category;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+}
