+++ src/main/java/com/wanted/codebombalms/domain/problems/set/exception/SetNotFoundException.java
@@ -1,0 +1,8 @@
+package com.wanted.codebombalms.domain.problems.set.exception;
+
+public class SetNotFoundException extends RuntimeException {
+
+    public SetNotFoundException(String message) {
+        super(message);
+    }
+}
