+++ src/main/java/com/wanted/codebombalms/domain/problems/submission/dto/request/SubmissionRequest.java
@@ -1,0 +1,15 @@
+package com.wanted.codebombalms.domain.problems.submission.dto.request;
+
+public class SubmissionRequest {
+
+    private Long userId;
+    private String submittedAnswer;
+
+    public Long getUserId() {
+        return userId;
+    }
+
+    public String getSubmittedAnswer() {
+        return submittedAnswer;
+    }
+}
